autobotserver/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.3
